justfile docs + rename to tech-mc-docs
