+++ tech-mc-docs/.gitignore
