+++ tech-mc-docs/babel.config.js
