+++ tech-mc-docs/blog/2021-08-01-mdx-blog-post.mdx
