+++ tech-mc-docs/docs/tutorial-extras/img/localeDropdown.png
