+++ tech-mc-docs/docusaurus.config.js
@@ -6,8 +6,8 @@ const darkCodeTheme = require('prism-react-renderer/themes/dracula');
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
-  title: 'My Site',
-  tagline: 'Dinosaurs are cool',
+  title: 'Technical Mc Docs',
+  tagline: 'Documentation on redstone technical mc that is used in this project! (Made so I don\'t forget stuff)',
   favicon: 'img/favicon.ico',
 
   // Set the production url of your site here
