+++ tech-mc-docs/package.json
