+++ tech-mc-docs/src/css/custom.css
