+++ tech-mc-docs/static/img/favicon.ico
